the OG library on Thingiverse was updated by another TV user
The update to the OG library added a lot of documentation (all internal). parameters for testing the modules of the library, and all the internal data additions to create a Configuration panel. This implies the creation of a .json file to hold the parameters, making them persistent
The OG and updated files where written in German which I translated to English to create my own version with expanded testing